Modification de la fonction qui envoie le(s) produit(s) au Local Storage + Ajout d'une condition qui vérifie la couleur et la quantité choisi

+++ front/js/product.js
@@ -3,7 +3,7 @@ const img = document.querySelector('.item__img');
 const title = document.getElementById('title');
 const price = document.getElementById('price');
 const description = document.getElementById('description');
-const selectColors = document.getElementById('colors');
+const color = document.getElementById('colors');
 const quantity = document.getElementById('quantity');
 const btnAddToCart = document.getElementById('addToCart');
 
@@ -12,11 +12,12 @@ const btnAddToCart = document.getElementById('addToCart');
 
 let url = window.location.href;
 let urlProduct = new URL(url);
-var search_params = new URLSearchParams(urlProduct.search); 
-var id = search_params.get('id');
+let search_params = new URLSearchParams(urlProduct.search); 
+let id = search_params.get('id');
 // ---------------------------------------------------------
 
 getProduct();
+addToCart();
 
 // ---------------------------------------------------------
 // fonction qui récupère les données de l'API => Json
@@ -61,47 +62,56 @@ function displayProduct(product) {
     // création de l'élément et affichage des otpions de couleurs dans le DOM
     for (let colors of product.colors) {
         let optionColors = document.createElement('option');
-        selectColors.appendChild(optionColors);
+        color.appendChild(optionColors);
         optionColors.value = colors;
         optionColors.innerHTML = colors;
     }
 };
 
 // ---------------------------------------------------------
+
 // fonction qui envoie le(s) produit(s) au Local Storage
+function addToCart(product){
 
-btnAddToCart.addEventListener('click', (e) => {
+    btnAddToCart.addEventListener('click', () => {
 
-    // variables contenants couleurs et quantité du produit
-    let selectedColor = selectColors.value;
-    let selectedQuantity = quantity.value;
+        // condition qui vérifie si une couleur est bien choisi et si la quantité > 0 et < 100
+        if (color.value != '' && quantity.value > 0 && quantity.value <= 100) {
+
+            // variables contenants couleurs et quantité du produit
+            let selectedColor = color.value;
+            let selectedQuantity = quantity.value;
+            
+            // objet contenant toutes les infos du produit selectionné
+            let selectedProduct = {
+                idItem : id,
+                nameItem : product.name,
+                colorItem : selectedColor,
+                quantityItem : selectedQuantity,
+                priceItem : product.price,
+                descritpionItem : product.description,
+                imgItem : product.imageUrl,
+                altImgItem : product.altTxt
+            };
+        
+            // variable contenant le Local Storage sours forme de tableau
+            localStorageProducts = [];
+        
+            // message de confirmation du panier
+            function confirmation() {
+                if(window.confirm(
+                    `Vous venez d'ajouter ${selectedQuantity} * ${product.name} ( ${selectedColor} ) dans votre panier !
+                    Pour consulter votre panier, cliquez sur OK`
+                )) {window.location.href ="cart.html";}
+            }
+        
+            // ajout de l'objet dans le Local Storage
+            localStorageProducts.push(selectedProduct);
+            localStorage.setItem('product', JSON.stringify(localStorageProducts));
+            console.table(localStorageProducts)
+            confirmation()
+        }
     
-    // objet contenant toutes les infos du produit selectionné
-    let selectedProduct = {
-        idProduct : id,
-        nameProduct : product.name,
-        colorProduct : selectedColor,
-        quantityProduct : selectedQuantity,
-        priceProduct : product.price,
-        descritpionProduct : product.description,
-        imgProduct : product.imageUrl,
-        altImgProduct : product.altTxt
-    };
+    });
+}
 
-    // variable contenant le Local Storage sours forme de tableau
-    localStorageProducts = [];
-
-    // message de confirmation du panier
-    function confirmation() {
-        if(window.confirm(
-            `Vous venez d'ajouter ${selectedQuantity} * ${product.name} ( ${selectedColor} ) dans votre panier !
-            Pour consulter votre panier, cliquez sur OK`
-        )) {window.location.href ="cart.html";}
-    }
-
-    // ajout de l'objet dans le Local Storage
-    localStorageProducts.push(selectedProduct);
-    localStorage.setItem('product', JSON.stringify(localStorageProducts));
-    console.table(localStorageProducts)
-    confirmation()
-});
